% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMRLOVA.R
\name{MVMRLOVA}
\alias{MVMRLOVA}
\title{MultiVariable Mendelian Randomization based on Latent Outcome Variable Approach(MVMRLOVA)}
\usage{
MVMRLOVA(
  betaY,
  betaX,
  betaYse,
  betaXse,
  ny,
  gwas_p,
  gwas_p2,
  permutn = 0,
  log_file = NULL
)
}
\arguments{
\item{betaY}{A numeric vector of standardized effect sizes (beta) from the outcome GWAS.}

\item{betaX}{A numeric matrix of standardized effect sizes (beta) from the exposures GWAS. Each column corresponds to a different exposure.}

\item{betaYse}{A numeric vector of standard errors for \code{betaY}.}

\item{betaXse}{A numeric matrix of standard errors for \code{betaX}.}

\item{gwas_p}{p-value threshold for the association between variants and the latent phenotype (the direct genetic effect). Default: 5e-2.}

\item{gwas_p2}{p-value threshold for exposure GWAS used in instrument variable selection. Default: 5e-8}

\item{permutn}{number of permutations, Default: 0.}

\item{log_file}{file name for saving the iterations to check convergence. Default: NULL.}
}
\value{
A list with the following elements:
\item{CausEst}{Estimate of causal effect.}
\item{CausEstSE}{Standard error of causal effect estimate.}
\item{CausEstP}{P-value from the z test for the causal effect.}
\item{IVs}{p-values for the direct causal effect of all Instrumental Variants (IV) on the outcome.}
\item{Valid}{Index of valid  instrumental variants.}
\item{sig_v}{the 5th percentile of the permuted p-values .}
\item{corrected_p-value}{corrected p-value based on the permutation distribution.}
}
\description{
Identifies vertical pleiotropy variants and estimates multivariable causal effects using GWAS summary statistics.
}
\details{
This function iteratively updates the set of vertical pleiotropy instruments (Valid) by identifying SNPs that pass significance criteria based on latent outcome variable GWAS statistics.
It assumes that both SNP effect sizes and the outcome trait are standardized.

The function uses the \code{MendelianRandomization} package to perform multivariable inverse-variance weighted (IVW) MR.
}
\keyword{Mendelian}
\keyword{randomization.}
