% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mvmrlova.R
\name{print.mvmrlova}
\alias{print.mvmrlova}
\title{Print method for MVMRLOVA result}
\usage{
\method{print}{mvmrlova}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mvmrlova}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for MVMRLOVA result
}
